python_ws/supervised_learning/requirements.txt
numpy
omegaconf
hydra-core
torch
tqdm
opencv-python
tensorboard